jest can run in both root and sub dirs
